pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillfortify"
version = "0.4.4"
description = "SkillFortify — Supply chain security scanner for AI agent skills. Supports 22 frameworks."
readme = "README.md"
requires-python = ">=3.11"
license = "Elastic-2.0"
authors = [
    { name = "Varun Pratap Bhardwaj", email = "varun.pratap.bhardwaj@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "cyclonedx-python-lib>=7.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/qualixar/skillfortify"
Repository = "https://github.com/qualixar/skillfortify"
Paper = "https://arxiv.org/abs/2603.00195"

[project.optional-dependencies]
sat = [
    "python-sat>=0.1.8.dev1",
]
registry = [
    "httpx>=0.27",
]
all = [
    "httpx>=0.27",
    "python-sat>=0.1.8.dev1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
    "ruff>=0.9",
]

[project.scripts]
skillfortify = "skillfortify.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/skillfortify"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".backup/",
    ".venv/",
    "venv/",
    ".claude/",
    "CLAUDE.md",
    "AGENTS.md",
    ".npmignore",
    ".code-review-graph/",
    ".gitnexus/",
    ".env*",
    "*.log",
    ".pytest_cache/",
    ".ruff_cache/",
    ".mypy_cache/",
    ".hypothesis/",
    "htmlcov/",
    "*.bak",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
